+++ iamf/cli/obu_processor_utils.h
@@ -5,15 +5,15 @@
 #include <list>
 #include <optional>
 
+#include "absl/base/nullability.h"
 #include "absl/status/statusor.h"
 #include "iamf/obu/mix_presentation.h"
-#include "iamf/obu/types.h"
 
 namespace iamf_tools {
 
 /*!\brief A selected Mix and Layout indices found in the Mixes. */
 struct SelectedMixPresentation {
-  DecodedUleb128 mix_presentation_id;
+  MixPresentationObu* absl_nonnull mix_presentation;
   Layout output_layout;
   // The index of the sub-mix where the selected Layout was found.
   int sub_mix_index;
@@ -39,6 +39,21 @@ absl::StatusOr<SelectedMixPresentation> FindMixPresentationAndLayout(
     const std::list<MixPresentationObu*>& supported_mix_presentations,
     const std::optional<Layout>& desired_layout,
     std::optional<uint32_t> desired_mix_presentation_id);
+
+/*!\brief Creates a simplified MixPresentation/Layout for rendering.
+ *
+ * The simplified MixPresentation will only have a single sub-mix and a single
+ * layout.
+ *
+ * \param mix_presentation MixPresentation to create a simplified version of.
+ * \param sub_mix_index The index of the sub-mix to keep.
+ * \param layout_index The index of the layout to keep.
+ * \return Simplified `MixPresentationObu`, or an error if the indices are out
+ *         of bounds.
+ */
+absl::StatusOr<MixPresentationObu> CreateSimplifiedMixPresentationForRendering(
+    const MixPresentationObu& mix_presentation, int sub_mix_index,
+    int layout_index);
 
 }  // namespace iamf_tools